% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biccalc2.R
\name{biccalc2}
\alias{biccalc2}
\title{BIC Calculator 2}
\usage{
biccalc2(pop_matrix_row, all_dataframe, Kvar, Kint, k)
}
\arguments{
\item{pop_matrix_row}{Vector containing a model}

\item{all_dataframe}{Dataframe with all variables}

\item{Kvar}{Maximum number of variables allowed in the final model}

\item{Kint}{Maximum number of interactions allowed in the final model}

\item{k}{Basis dimension for nonparametric terms estimated using cubic splines.}
}
\value{
BIC value of the model
}
\description{
Alternative function to calculate the BIC of the model. Internal use only.
}
