% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gagam.R
\name{gagam}
\alias{gagam}
\title{Genetic Algorithm for Generalized Additive Models}
\usage{
gagam(
  y,
  x,
  pop_size = 500,
  Kvar = 15,
  Kint = 0,
  no_gen = 100,
  p_m = 0.05,
  p_int = 0.1,
  p_nonpar = 0.1,
  p_int_nonpar = 0.1,
  multicore = TRUE,
  k = 10
)
}
\arguments{
\item{y}{Vector, matrix, or data frame containing observations of the dependent variable.}

\item{x}{Matrix or data frame containing all considered explanatory variables.}

\item{pop_size}{Size of the population (in multiples of 500). Default is 500.}

\item{Kvar}{Maximum number of variables allowed in the final model. Default is 15.}

\item{Kint}{Maximum number of interactions allowed in the final model. Default is 0.}

\item{no_gen}{Number of generations until convergence. Default is 100.}

\item{p_m}{Mutation rate for variables. Default is 0.05.}

\item{p_int}{Mutation rate of interactions of variables. Default is 0.1.}

\item{p_nonpar}{Mutation rate for the linear/nonparametric indicators for variables. Default is 0.1.}

\item{p_int_nonpar}{Mutation rate for the linear/nonparametric indicators for interactions. Default is 0.1.}

\item{multicore}{Whether to use multiple cores in computation. Strongly recommended but may not work on Windows. Default is TRUE.}

\item{k}{Basis dimension for nonparametric terms estimated using cubic splines. Default is 10.}
}
\value{
A list containing: gam object from the mgcv package (estimated best model), vector of indexes of variables included linearly, vector of indexes of variables included nonparametrically (and the same lists for interactions if Kint is greater than 0).
}
\description{
Implements the genetic algorithm for simultaneous variable selection and structure discovery in generalized additive models. For a given dependent variable and a set of explanatory variables, the genetic algorithm determines which regressors should be included linearly, which nonparametrically, and which should be excluded from the regression equation. The aim is to minimize the Bayesian Information Criterion value of the model.
}
\examples{
N <- 500
set.seed(123)
xdat <- matrix(rnorm(N*10,0,1),nrow=N,ncol=10)
ydat <- 4*xdat[,1]+5*xdat[,2]+6*xdat[,3]+(xdat[,4])^2 + 4 + rnorm(N,0,0.25)

\dontrun{
example_gagam <- gagam(ydat,xdat,Kvar = 6,no_gen = 50)
}

}
