% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ints_names_func.R
\name{ints_names_func}
\alias{ints_names_func}
\title{Extract Interaction Names}
\usage{
ints_names_func(ints, xcolnames)
}
\arguments{
\item{ints}{Vector like temp_extract_vars$linear_ints}

\item{xcolnames}{Vector of column names of matrix x.}
}
\value{
Vector of correctly named interactions
}
\description{
Names the interactions in the output correctly. Internal use only.
}
\keyword{internal}
