% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_func.R
\name{mut_func}
\alias{mut_func}
\title{Model Mutation}
\usage{
mut_func(mutation_row, Kvar, Kint, pm, pint, pnonpar, pintnonpar, regressors)
}
\arguments{
\item{mutation_row}{Vector containing a model.}

\item{Kvar}{Maximum number of variables allowed in the final model}

\item{Kint}{Maximum number of interactions allowed in the final model}

\item{pm}{Mutation rate for variables.}

\item{pint}{Mutation rate for interactions of variables.}

\item{pnonpar}{Mutation rate for model structure of main terms.}

\item{pintnonpar}{Mutation rate for model structure of interaction terms.}

\item{regressors}{Total number of variables considered.}
}
\value{
Vector containing the mutated model.
}
\description{
Mutates both the variables included and the model structure. Internal use only.
}
\examples{
mutation_row <- c("x1","0","x4","x3","0","x8","1","0","0","0","0","1")
Kvar <- 10
Kint <- 0
pm <- 0.05
pint <- 0.1
pnonpar <- 0.1
pintnonpar <- 0.1
regressors <- 20

\dontrun{
mut_func(mutation_row,Kvar,Kint,pm,pint,pnonpar,pintnonpar,regressors)
}

}
