% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparmut_func.R
\name{nonparmut_func}
\alias{nonparmut_func}
\title{Mutate Model Structure}
\usage{
nonparmut_func(mutation_row, Kvar, Kint, pnonpar, pintnonpar)
}
\arguments{
\item{mutation_row}{Vector containing a model.}

\item{Kvar}{Maximum number of variables allowed in the final model}

\item{Kint}{Maximum number of interactions allowed in the final model}

\item{pnonpar}{Mutation rate for model structure of main terms.}

\item{pintnonpar}{Mutation rate for model structure of interaction terms.}
}
\value{
Vector containing the mutated model.
}
\description{
Mutates the model structure, leaving the included variables unchanged. Internal use only.
}
\examples{
mutation_row <- c("x1","0","x4","x3","0","x8","1","0","0","0","0","1")
Kvar <- 10
Kint <- 0
pnonpar <- 0.1
pintnonpar <- 0.1

\dontrun{
nonparmut_func(mutation_row,Kvar,Kint,pnonpar,pintnonpar)
}

}
