% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gagam.R
\name{plot.gagam}
\alias{plot.gagam}
\title{Plot GAGAM}
\usage{
\method{plot}{gagam}(x, ..., reduc = NULL)
}
\arguments{
\item{x}{Fitted \code{gagam} object}

\item{...}{Any other arguments to pass to \code{\link[mgcv]{plot.gam}}.}

\item{reduc}{If you wish to summarize one of the reduced models, specify the index of reduction. Specify only one index at a time. Default is NULL.}
}
\description{
Plot method for \code{gagam} objects.
}
