% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.gagam.R
\name{predict.gagam}
\alias{predict.gagam}
\title{Prediction from GAGAM}
\usage{
\method{predict}{gagam}(object, ..., newdata, reduc = NULL)
}
\arguments{
\item{object}{\code{gagam} object (result of the \code{gagam()} function)}

\item{...}{Any other arguments to pass to \code{\link[mgcv]{predict.gam}}.}

\item{newdata}{Matrix or data frame with explanatory variables. The order of columns and their names must be the same as that used to construct the gagam object. If missing, the function extracts in-sample predictions.}

\item{reduc}{If prediction should be from one of the reduced models, specify the index of reduction. Specify only one index at a time. Default is NULL.}
}
\value{
A vector of predictions.
}
\description{
Predict method for \code{gagam} objects.
}
\examples{
N <- 1000
set.seed(123)
xdat2 <- matrix(rnorm(N*10,0,1),nrow=N,ncol=10)
ydat2 <- 4*xdat2[,1]+5*xdat2[,2]+6*xdat2[,3]+(xdat2[,4])^2 + 4 + rnorm(N,0,0.25)

xdat <- xdat2[1:500,]
ydat <- ydat2[1:500]

\dontrun{
example_gagam <- gagam(ydat,xdat,Kvar = 6,no_gen = 50)
}

xdat <- xdat2[501:1000,]
ydat <- ydat2[501:1000]

newdata <- as.data.frame(cbind(as.matrix(ydat),xdat))

\dontrun{
predict_gagam(example_gagam,newdata)
}

}
