% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gagam.R
\name{predict_gagam}
\alias{predict_gagam}
\title{Prediction from GAGAM}
\usage{
predict_gagam(gagam, newdata, reduc = NULL)
}
\arguments{
\item{gagam}{\code{gagam} object (result of the \code{gagam()} function)}

\item{newdata}{Matrix or data frame with explanatory variables. The order of columns must be the same as that used to construct the gagam object. If missing, the function extracts in-sample predictions.}

\item{reduc}{If prediction should be from one of the reduced models, specify the index of reduction. Specify only one index at a time. Default is NULL.}
}
\value{
A vector of predictions.
}
\description{
Takes the result of variable and structure discovery search using a genetic algorithm (\code{gagam()} function) and produces predictions given a set of values for the explanatory variables.
}
\examples{
N <- 1000
set.seed(123)
xdat2 <- matrix(rnorm(N*10,0,1),nrow=N,ncol=10)
ydat2 <- 4*xdat2[,1]+5*xdat2[,2]+6*xdat2[,3]+(xdat2[,4])^2 + 4 + rnorm(N,0,0.25)

xdat <- xdat2[1:500,]
ydat <- ydat2[1:500]

\dontrun{
example_gagam <- gagam(ydat,xdat,Kvar = 6,no_gen = 50)
}

xdat <- xdat2[501:1000,]
ydat <- ydat2[501:1000]

newdata <- as.data.frame(cbind(as.matrix(ydat),xdat))

\dontrun{
predict_gagam(example_gagam,newdata)
}

}
