% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonpar_func.R
\name{remove_nonpar_func}
\alias{remove_nonpar_func}
\title{Reduction of Nonparametric Terms}
\usage{
remove_nonpar_func(
  best_ind,
  all_dataframe,
  Kvar,
  Kint,
  k,
  family,
  method,
  optimizer
)
}
\arguments{
\item{best_ind}{Vector contain the model.}

\item{all_dataframe}{Data frame with all variables.}

\item{Kvar}{Maximum number of variables allowed in the final model}

\item{Kint}{Maximum number of interactions allowed in the final model}

\item{k}{Basis dimension for nonparametric terms estimated using cubic splines.}

\item{family}{Specifies the family for the gam (see ?family and ?family.mgcv). Default is gaussian().}

\item{method}{Specifies the metric for smoothing parameter selection (see ?gam). Default is "REML".}

\item{optimizer}{Specifies the numerical optimization algorithm for the gam (see ?gam). Default is c("outer","newton").}
}
\value{
Reduced model.
}
\description{
Additional reduction of nonparametric terms after the genetic algorithm has finished. Internal use only.
}
