% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_func.R
\name{sorting_func}
\alias{sorting_func}
\title{Model Sort}
\usage{
sorting_func(individual_row, Kvar, Kint)
}
\arguments{
\item{individual_row}{Vector containing the model}

\item{Kvar}{Maximum number of variables allowed in the final model}

\item{Kint}{Maximum number of interactions allowed in the final model}
}
\value{
A string summarizing the model.
}
\description{
Sorts the variables in the model and summarizes the variables included and whether they enter linearly or not in a string. For internal use in memoization.
}
\examples{
individual_row <- c("x1","0","x4","x3","0","x8","1","0","0","0","0","1")
Kvar <- 6
Kint <- 0

\dontrun{
sorting_func(individual_row,Kvar,Kint)
}

}
\keyword{internal}
