% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gagam.R
\name{summary.gagam}
\alias{summary.gagam}
\title{Summarize GAGAM}
\usage{
\method{summary}{gagam}(object, ..., reduc = NULL)
}
\arguments{
\item{object}{Fitted \code{gagam} object}

\item{...}{Any other arguments to pass to \code{\link[mgcv]{summary.gam}}.}

\item{reduc}{If you wish to summarize one of the reduced models, specify the index of reduction. Specify only one index at a time. Default is NULL.}
}
\value{
Summary of the \code{gam} object.
}
\description{
Summary method for \code{gagam} objects.
}
